[FIX] Removed useless functions and variables
