+++ include/types.h
@@ -66,8 +66,6 @@ typedef struct environment {
     struct environment *next;
 } env_t;
 
-typedef bool logical_checker_t(command_t *cmd, char *const *array, \
-    const env_t *env);
 typedef bool redirection_checker_t(command_t *cmd, char *const *array,
     const env_t *env);
 typedef void redirection_function_t(env_t **env, char *const *args);
@@ -81,7 +79,6 @@ typedef struct redirection {
 
 typedef struct logical {
     char *type;
-    logical_checker_t *checker;
     separatour_next_type_t separator;
 } logical_t;
