+++ src/command/logical.c
@@ -5,21 +5,59 @@
 ** Logical function checkers
 */
 
+#include <stddef.h>
+#include <stdio.h>
+#include <string.h>
 #include "logical.h"
+#include "messages.h"
 
-bool check_and(command_t *cmd, char *const *array, const env_t *env)
+static bool check_logical(command_t *command, bool empty)
 {
+    command_t *prev = command->prev;
+    separatour_next_type_t next = command->separator_next;
+
+    if (empty && next == OR) {
+        fprintf(stderr, "%s\n", MISSING_COMMAND);
+        return (false);
+    }
+    if (empty && prev != NULL) {
+        next = prev->separator_next;
+        if (next == OR ) {
+            fprintf(stderr, "%s\n", MISSING_COMMAND);
+            return (false);
+        }
+    }
     return (true);
 }
 
-bool check_or(command_t *cmd, char *const *array, const env_t *env)
+bool check_logicals(command_t *list, shell_t *shell)
 {
+    bool empty = false;
+    command_t *current = list;
+
+    while (current != NULL) {
+        empty = (current->args[0] == NULL || strlen(current->args[0]) == 0);
+        if (!check_logical(current, empty)) {
+            shell->ret = 1;
+            return (false);
+        }
+        current = current->next;
+    }
     return (true);
 }
 
-bool check_semicolon(command_t *cmd, char *const *array, const env_t *env)
+bool should_ignore(command_t *command)
 {
-    return (true);
+    if (command->prev == NULL) {
+        return (false);
+    }
+    if (command->prev->separator_next == AND && command->prev->ret != 0) {
+        return (true);
+    }
+    if (command->prev->separator_next == OR && command->prev->ret == 0) {
+        return (true);
+    }
+    return (false);
 }
 
 void apply_logical(command_t *command, char const *separator)
